requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
requests==2.28.1
loguru==0.6.0
peewee==3.15.2
